backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Core ML dependencies
torch>=2.3.1
torchvision>=0.18.1
numpy>=1.24.0
opencv-contrib-python>=4.9.0  # Includes legacy trackers (CSRT)
Pillow>=10.0.0

# YOLOv10 - using ultralytics
ultralytics==8.1.0

# CLIP
git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm

# OSTrack dependencies
timm>=1.0.0
einops==0.7.0

# Utility
python-dotenv==1.0.0

# Python Desktop App Dependencies
google-genai>=0.2.0  # Gemini Live API
pyaudio>=0.2.13  # Audio capture/playback
PyQt6>=6.6.0  # GUI framework
websockets>=12.0  # WebSocket client for bridge
